Add timeout in FSM (#17)
This PR is introducing a new configurable parameter,
voice_calls.max_duration that acts as a safety mechanism in case the
VOIP addon is getting out of sync with the Baresip instance.
In particular it imposes a maximum duration on any voice call, whether
it's being estabilished (ringing) or has already been established.

This PR also improves logging.

---------

Co-authored-by: Francesco Montorsi <francesco.montorsi@eu.equinix.com>